+++ tests/OpenVpnTest.php
@@ -37,10 +37,10 @@ class OpenVpnTest extends TestCase
 
     public function testWriteProfiles()
     {
-        $this->openVpn->writeProfiles($this->serverClient, 'default', 'openvpn', 'openvpn', true);
+        $this->openVpn->writeProfiles($this->serverClient, 'openvpn', 'openvpn', true);
         $this->assertSame(
-            trim(file_get_contents(sprintf('%s/default-internet-0.conf', $this->tmpDir))),
-            trim(file_get_contents(sprintf('%s/data/default-internet-0.conf', __DIR__)))
+            trim(file_get_contents(sprintf('%s/internet-0.conf', $this->tmpDir))),
+            trim(file_get_contents(sprintf('%s/data/internet-0.conf', __DIR__)))
         );
     }
 }